first commit8

+++ src/tts20.h
@@ -16,16 +16,13 @@ namespace em {
 
 /**
  * @~Chinese
- * @brief TTS20 文本转语音模块驱动类
- * @details 该类提供了与TTS20文本转语音模块通信的接口，支持通过I2C协议
- *          发送文本、控制播放状态、查询设备信息等功能
+ * @class Tts20
+ * @brief Tts20是用于TTS20模块的驱动类，用于语音合成。
  */
 /**
  * @~English
- * @brief TTS20 Text-to-Speech Module Driver Class
- * @details This class provides the interface for communicating with the TTS20
- *          text-to-speech module, supporting functions such as sending text via
- *          I2C protocol, controlling playback status, and querying device information
+ * @class Tts20
+ * @brief Tts20 is a driver class for TTS20 module, used for speech synthesis。
  */
 class Tts20 {
  public:
@@ -102,18 +99,14 @@ class Tts20 {
   /**
    * @~Chinese
    * @brief 文本转语音并播放。
-   * @param[in] text 要播放的文本数据，数据长度不大于4K个字节。
-   * @return 是否成功合成并开始播放。
-   * @details 文本长度限制：单次合成文本量最多4K字节。如果文本超过此限制，播放将失败。
-   *          - 建议在发送前检查文本长度，避免因文本过长导致的播放失败。
+   * @param[in] text 要播放的文本数据。
+   * @return 是否成功播放。
    */
   /**
    * @~English
    * @brief Convert text to speech and play.
-   * @param[in] text The text data to be played should not exceed a length of 4K bytes.
+   * @param[in] text The text data to be played.
    * @return Has it been successfully synthesized and started playing.
-   * @details Text length limit: Maximum 4K bytes per synthesis. Playback will fail if text exceeds this limit.
-   *          - It is recommended to check text length before sending to avoid playback failure due to excessive text length.
    */
   bool Play(const String &text);
 
@@ -171,7 +164,7 @@ class Tts20 {
 
   void ClearRxBuffer();
   void I2cWrite(const uint8_t *data, const uint16_t size);
-  uint16_t I2cRead(uint8_t *buffer, const uint16_t expected_length, const uint32_t timeout_ms);
+  uint8_t I2cRead(uint8_t *buffer, const uint8_t expected_length, const uint32_t timeout_ms);
   bool ReadUntil(const uint8_t target_byte, const uint32_t timeout_ms);
 
   const uint8_t i2c_address_ = kDefaultI2cAddress;
